0.9.2: Implement Simple Mode for joystick usage
Introduced a new default "Simple Mode" which lacks filtering and only
supports downloading a single file at a time. But, it can be operated
entirely without a keyboard.

Normal mode would also work with a joystick, if a button were mapped to
spacebar. For some reason this isn't done, and while it would be
possible to read joystick button presses, it involves parsing a lot of
device-specific data, which is undesirable.