0.9.3: Local CHD file organising implemented
Since all CHD files are put into their own directories, extended
same logic to be usable also locally. Currently a somewhat hidden
feature, it is summoned from the command line by passing path
to the directory to be sorted as an argument:

% ./romweasel.sh /media/fat/games/PSX

Tested with full PSX library, out of 414 multi-CD games, 57 fail
to automatically compute a unique location and most of these are
unsolvable without resorting to hard-coding.